pyproject.toml
[tool.poetry]
name = "django-perms"
version = "0.1.0"
description = ""
authors = ["xolmomin <xolmomin@gmail.com>"]
readme = "README.md"
packages = [{include = "django_perms"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.6"
drf-yasg = "^1.21.4"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
